feat: Refactor data handling and type casting in several components
In this commit, the way data is managed and cast has been updated in several components. Both 'ProjectsSection' and 'SkillsSection' now cast incoming data to their specific structures. Moreover, the 'useFetch' hook now provides a more specific data type, allowing received data to be either an object, an array of objects, or null. Finally, to ensure icons are displayed correctly, a more specific type was used for the SkillIcon in 'icon-lib.tsx'.